Better plan forwards, should've finished game tomorrow

+++ game-project/powerups/types/powerup_drop.gd
@@ -3,5 +3,7 @@ class_name PowerupDrop
 
 func execute(pe:ModulePowerupExecuter, _dt:float) -> bool:
 	for player in pe.players_here:
+		if not player.parasol_grabber.has_parasol(): continue
 		player.parasol_grabber.drop()
+		GSignal.feedback.emit(pe.global_position, "Dropped!")
 	return true
